feat: add cover images to search results (#167)
* move docker file and update user permissions

* Fix bug with match custom search modal not closing

* Prevent failed book from blocking others from starting to process

When matching books to ASINs if one or more fails it will still start processing others and just show error for those that failed.

* message warning to error

* Remove `rootdir` in match view since the full path is passed in the session already

* Add env var for celery worker count

* Updates to the Read me

* fix: Add error catch for null input_data in merge task

* fix: update src_path on book model for books already in the db

* docs: forgot to save book model after updating the src path

* feat: add x to remove item from search selection

adds a modal to check if you want to remove the item from the search selection

* Feat: Add cover image to search results and book view

doc: Add URL field to book model for the cover image. This will be used in the book view to show cover images
added a data migration to add the url field to books and fetch the image url to populate for books already in the db
added a cover not available default image to static files
removed or updated some logging, especially in the js files

* doc: Move Dockerfile back into docker folder

Updated the dockerfile to only copy in what is needed and where to put it them
added a docker compose example in the readme

* Fix: fixed the search drop down full width size

Fixed a regression in the search dropdown where they were not full width anymore
Fixed where the button submit button would be greyed out after removing an item with no asin value found

* Feat: Add cover image to search results and book view

doc: Add URL field to book model for the cover image. This will be used in the book view to show cover images
added a data migration to add the url field to books and fetch the image url to populate for books already in the db
added a cover not available default image to static files
removed or updated some logging, especially in the js files

* feat: add x to remove item from search selection

adds a modal to check if you want to remove the item from the search selection

* Fix: fixed the search drop down full width size

Fixed a regression in the search dropdown where they were not full width anymore
Fixed where the button submit button would be greyed out after removing an item with no asin value found

* Feat: Add cover image to search results and book view

doc: Add URL field to book model for the cover image. This will be used in the book view to show cover images
added a data migration to add the url field to books and fetch the image url to populate for books already in the db
added a cover not available default image to static files
removed or updated some logging, especially in the js files

* fix sonar analysis code errors

* clean up fetchOption js method

---------

Co-authored-by: David Dembeck <71412966+djdembeck@users.noreply.github.com>